pyproject.toml
[project]
name = "queens-puzzle-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=11.1.0",
]
